chore(): refac and add utils
