+++ packages/analytics/package.json
@@ -3,10 +3,10 @@
   "version": "0.0.2",
   "description": "",
   "scripts": {
-    "clean": "rm -rf dist",
     "test": "echo \"Error: no test specified\" && exit 0",
-    "build": "npm run clean && rm -f *.tsbuildinfo && tsc",
-    "prepare": "npm run build"
+    "build": "npm run build:clean && npm run build:prepare",
+    "build:clean": "rm -rf dist && rm -f *.tsbuildinfo",
+    "build:prepare": "tsc"
   },  
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
