+++ packages/analytics/src/const.ts
@@ -1,0 +1,4 @@
+const DEFAULT_SCOPES = ["analytics:website:tracking"];
+const SERVICE_NAME = "analytics";
+
+export { DEFAULT_SCOPES, SERVICE_NAME };
