+++ packages/analytics/src/index.ts
@@ -1,1 +1,5 @@
-export * from "./components/useAnalyticsTracker";
+export * from "./useAnalyticsTracker";
+export * from "./trackNavigationEvent";
+export * from "./types";
+export * from "./const";
+export * from "./client";
